extract CreateDropDown
